Put the translation structure in new file
